+++ pypesto/objective/finite_difference.py
@@ -327,6 +327,7 @@ class FD(ObjectiveBase):
         self.delta_grad: FDDelta = to_delta(delta_grad)
         self.delta_res: FDDelta = to_delta(delta_res)
         self.method: str = method
+        self.pre_post_processor = obj.pre_post_processor
 
         if method not in FD.METHODS:
             raise ValueError(
@@ -373,6 +374,7 @@ class FD(ObjectiveBase):
         x: np.ndarray,
         sensi_orders: tuple[int, ...],
         mode: ModeType,
+        return_dict: bool,
         **kwargs,
     ) -> ResultDict:
         """
@@ -383,11 +385,17 @@ class FD(ObjectiveBase):
         """
         if mode == MODE_FUN:
             result = self._call_mode_fun(
-                x=x, sensi_orders=sensi_orders, **kwargs
+                x=x,
+                sensi_orders=sensi_orders,
+                return_dict=return_dict,
+                **kwargs,
             )
         elif mode == MODE_RES:
             result = self._call_mode_res(
-                x=x, sensi_orders=sensi_orders, **kwargs
+                x=x,
+                sensi_orders=sensi_orders,
+                return_dict=return_dict,
+                **kwargs,
             )
         else:
             raise ValueError("This mode is not supported.")
@@ -398,6 +406,7 @@ class FD(ObjectiveBase):
         self,
         x: np.ndarray,
         sensi_orders: tuple[int, ...],
+        return_dict: bool,
         **kwargs,
     ) -> ResultDict:
         """Handle calls in function value mode.
@@ -408,6 +417,7 @@ class FD(ObjectiveBase):
         sensi_orders_obj, result = self._call_from_obj_fun(
             x=x,
             sensi_orders=sensi_orders,
+            return_dict=return_dict,
             **kwargs,
         )
 
@@ -429,13 +439,21 @@ class FD(ObjectiveBase):
         def f_fval(x):
             """Short-hand to get a function value."""
             return self.obj.call_unprocessed(
-                x=x, sensi_orders=(0,), mode=MODE_FUN, **kwargs
+                x=x,
+                sensi_orders=(0,),
+                mode=MODE_FUN,
+                return_dict=return_dict,
+                **kwargs,
             )[FVAL]
 
         def f_grad(x):
             """Short-hand to get a gradient value."""
             return self.obj.call_unprocessed(
-                x=x, sensi_orders=(1,), mode=MODE_FUN, **kwargs
+                x=x,
+                sensi_orders=(1,),
+                mode=MODE_FUN,
+                return_dict=return_dict,
+                **kwargs,
             )[GRAD]
 
         # update delta vectors
@@ -487,6 +505,7 @@ class FD(ObjectiveBase):
         self,
         x: np.ndarray,
         sensi_orders: tuple[int, ...],
+        return_dict: bool,
         **kwargs,
     ) -> ResultDict:
         """Handle calls in residual mode.
@@ -497,6 +516,7 @@ class FD(ObjectiveBase):
         sensi_orders_obj, result = self._call_from_obj_res(
             x=x,
             sensi_orders=sensi_orders,
+            return_dict=return_dict,
             **kwargs,
         )
 
@@ -507,7 +527,11 @@ class FD(ObjectiveBase):
         def f_res(x):
             """Short-hand to get a function value."""
             return self.obj.call_unprocessed(
-                x=x, sensi_orders=(0,), mode=MODE_RES, **kwargs
+                x=x,
+                sensi_orders=(0,),
+                mode=MODE_RES,
+                return_dict=return_dict,
+                **kwargs,
             )[RES]
 
         # update delta vector
@@ -532,6 +556,7 @@ class FD(ObjectiveBase):
         self,
         x: np.ndarray,
         sensi_orders: tuple[int, ...],
+        return_dict: bool,
         **kwargs,
     ) -> tuple[tuple[int, ...], ResultDict]:
         """
@@ -553,7 +578,11 @@ class FD(ObjectiveBase):
         result = {}
         if sensi_orders_obj:
             result = self.obj.call_unprocessed(
-                x=x, sensi_orders=sensi_orders_obj, mode=MODE_FUN, **kwargs
+                x=x,
+                sensi_orders=sensi_orders_obj,
+                mode=MODE_FUN,
+                return_dict=return_dict,
+                **kwargs,
             )
         return sensi_orders_obj, result
 
@@ -561,6 +590,7 @@ class FD(ObjectiveBase):
         self,
         x: np.ndarray,
         sensi_orders: tuple[int, ...],
+        return_dict: bool,
         **kwargs,
     ) -> tuple[tuple[int, ...], ResultDict]:
         """
@@ -580,7 +610,11 @@ class FD(ObjectiveBase):
         result = {}
         if sensi_orders_obj:
             result = self.obj.call_unprocessed(
-                x=x, sensi_orders=sensi_orders_obj, mode=MODE_RES, **kwargs
+                x=x,
+                sensi_orders=sensi_orders_obj,
+                mode=MODE_RES,
+                return_dict=return_dict,
+                **kwargs,
             )
         return sensi_orders_obj, result