+++ pypesto/objective/priors.py
@@ -1,9 +1,9 @@
 import logging
 import math
 from collections.abc import Sequence
-from copy import deepcopy
 from typing import Callable, Union
 
+import cloudpickle
 import numpy as np
 
 from .. import C
@@ -68,16 +68,20 @@ class NegLogParameterPriors(ObjectiveBase):
         self.prior_list = prior_list
         super().__init__(x_names)
 
-    def __deepcopy__(self, memodict=None):
-        """Create deepcopy of object."""
-        other = NegLogParameterPriors(deepcopy(self.prior_list))
-        return other
+    def __getstate__(self):
+        """Get state using cloudpickle."""
+        return cloudpickle.dumps(self.__dict__)
+
+    def __setstate__(self, state):
+        """Set state using cloudpickle."""
+        self.__dict__.update(cloudpickle.loads(state))
 
     def call_unprocessed(
         self,
         x: np.ndarray,
         sensi_orders: tuple[int, ...],
         mode: C.ModeType,
+        return_dict: bool,
         **kwargs,
     ) -> ResultDict:
         """