+++ pypesto/objective/roadrunner/petab_importer_roadrunner.py
@@ -6,24 +6,28 @@ depends on the noise model specified in the provided PEtab problem.
 """
 from __future__ import annotations
 
+import logging
 import numbers
 import re
+import warnings
 from collections.abc import Iterable
 from pathlib import Path
 from typing import Any
 
-import libsbml
-import petab
-import roadrunner
-from petab.C import (
-    OBSERVABLE_FORMULA,
-    PREEQUILIBRATION_CONDITION_ID,
-    SIMULATION_CONDITION_ID,
-)
-from petab.models.sbml_model import SbmlModel
-from petab.parameter_mapping import ParMappingDictQuadruple
+try:
+    import petab.v1 as petab
+    from petab.v1.C import (
+        OBSERVABLE_FORMULA,
+        PREEQUILIBRATION_CONDITION_ID,
+        SIMULATION_CONDITION_ID,
+    )
+    from petab.v1.models.sbml_model import SbmlModel
+    from petab.v1.parameter_mapping import ParMappingDictQuadruple
+except ImportError:
+    petab = None
 
-from ...petab.importer import PetabStartpoints
+import pypesto.C
+
 from ...problem import Problem
 from ...startpoint import StartpointMethod
 from ..aggregated import AggregatedObjective
@@ -31,6 +35,15 @@ from ..priors import NegLogParameterPriors, get_parameter_prior_dict
 from .road_runner import RoadRunnerObjective
 from .roadrunner_calculator import RoadRunnerCalculator
 from .utils import ExpData
+
+try:
+    import libsbml
+    import roadrunner
+except ImportError:
+    roadrunner = None
+    libsbml = None
+
+logger = logging.getLogger(__name__)
 
 
 class PetabImporterRR:
@@ -57,6 +70,14 @@ class PetabImporterRR:
         validate_petab:
             Flag indicating if the PEtab problem shall be validated.
         """
+        warnings.warn(
+            "The RoadRunner importer is deprecated and will be removed in "
+            "future versions. Please use the generic PetabImporter instead "
+            "with `simulator_type='roadrunner'`. Everything else will stay "
+            "same.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
         self.petab_problem = petab_problem
         if validate_petab:
             if petab.lint_problem(petab_problem):
@@ -208,15 +229,41 @@ class PetabImporterRR:
                 preeq_id = condition.get(PREEQUILIBRATION_CONDITION_ID)
                 sim_id = condition.get(SIMULATION_CONDITION_ID)
                 if preeq_id:
-                    mapping_per_condition[0][
-                        override
-                    ] = self.petab_problem.condition_df.loc[preeq_id, override]
-                    mapping_per_condition[2][override] = "lin"
+                    parameter_id_or_value = (
+                        self.petab_problem.condition_df.loc[preeq_id, override]
+                    )
+                    mapping_per_condition[0][override] = parameter_id_or_value
+                    if isinstance(parameter_id_or_value, str):
+                        mapping_per_condition[2][
+                            override
+                        ] = self.petab_problem.parameter_df.loc[
+                            parameter_id_or_value, petab.PARAMETER_SCALE
+                        ]
+                    elif isinstance(parameter_id_or_value, numbers.Number):
+                        mapping_per_condition[2][override] = pypesto.C.LIN
+                    else:
+                        raise ValueError(
+                            "The parameter value in the condition table "
+                            "is not a number or a parameter ID."
+                        )
                 if sim_id:
-                    mapping_per_condition[1][
-                        override
-                    ] = self.petab_problem.condition_df.loc[sim_id, override]
-                    mapping_per_condition[3][override] = "lin"
+                    parameter_id_or_value = (
+                        self.petab_problem.condition_df.loc[sim_id, override]
+                    )
+                    mapping_per_condition[1][override] = parameter_id_or_value
+                    if isinstance(parameter_id_or_value, str):
+                        mapping_per_condition[3][
+                            override
+                        ] = self.petab_problem.parameter_df.loc[
+                            parameter_id_or_value, petab.PARAMETER_SCALE
+                        ]
+                    elif isinstance(parameter_id_or_value, numbers.Number):
+                        mapping_per_condition[3][override] = pypesto.C.LIN
+                    else:
+                        raise ValueError(
+                            "The parameter value in the condition table "
+                            "is not a number or a parameter ID."
+                        )
         return mapping
 
     def create_objective(
@@ -257,6 +304,7 @@ class PetabImporterRR:
             petab_problem=self.petab_problem,
             calculator=calculator,
             x_names=x_names,
+            x_ids=x_names,
         )
 
     def create_prior(self) -> NegLogParameterPriors | None:
@@ -279,6 +327,13 @@ class PetabImporterRR:
                 isinstance(prior_type_entry, str)
                 and prior_type_entry != petab.PARAMETER_SCALE_UNIFORM
             ):
+                # check if parameter for which prior is defined is a fixed parameter
+                if x_id in self.petab_problem.x_fixed_ids:
+                    logger.warning(
+                        f"Parameter {x_id} is marked as fixed but has a "
+                        f"prior defined. This might be unintended."
+                    )
+
                 prior_params = [
                     float(param)
                     for param in self.petab_problem.parameter_df.loc[
@@ -306,6 +361,8 @@ class PetabImporterRR:
             Additional keyword arguments passed on to
             :meth:`pypesto.startpoint.FunctionStartpoints.__init__`.
         """
+        from ...petab.util import PetabStartpoints
+
         return PetabStartpoints(petab_problem=self.petab_problem, **kwargs)
 
     def create_problem(