+++ pypesto/objective/roadrunner/roadrunner_calculator.py
@@ -2,14 +2,12 @@
 
 Handles all RoadRunner.simulate calls, calculates likelihoods and residuals.
 """
+from __future__ import annotations
+
 import numbers
 from collections.abc import Sequence
-from typing import Optional
 
 import numpy as np
-import petab
-import roadrunner
-from petab.parameter_mapping import ParMappingDictQuadruple
 
 from ...C import (
     FVAL,
@@ -27,6 +25,16 @@ from .utils import (
     simulation_to_measurement_df,
     unscale_parameters,
 )
+
+try:
+    import petab.v1 as petab
+    from petab.v1.parameter_mapping import ParMappingDictQuadruple
+except ImportError:
+    petab = None
+try:
+    import roadrunner
+except ImportError:
+    roadrunner = None
 
 LLH_TYPES = {
     "lin_normal": lambda measurement, simulation, sigma: -0.5
@@ -69,7 +77,7 @@ class RoadRunnerCalculator:
         x_ids: Sequence[str],
         parameter_mapping: list[ParMappingDictQuadruple],
         petab_problem: petab.Problem,
-        solver_options: Optional[SolverOptions],
+        solver_options: SolverOptions | None = None,
     ):
         """Perform the RoadRunner call and obtain objective function values.
 
@@ -240,10 +248,10 @@ class RoadRunnerCalculator:
     def fill_in_parameters(
         self,
         problem_parameters: dict,
-        roadrunner_instance: Optional[roadrunner.RoadRunner] = None,
-        parameter_mapping: Optional[ParMappingDictQuadruple] = None,
+        roadrunner_instance: roadrunner.RoadRunner | None = None,
+        parameter_mapping: ParMappingDictQuadruple | None = None,
         preeq: bool = False,
-        filling_mode: Optional[str] = None,
+        filling_mode: str | None = None,
     ) -> dict:
         """Fill in parameters into the roadrunner instance.