+++ pypesto/result/__init__.py
@@ -7,7 +7,7 @@ the pypesto components. It contains sub-results for
 optimization, profiling, sampling.
 """
 
-from .optimize import OptimizeResult, OptimizerResult
+from .optimize import LazyOptimizerResult, OptimizeResult, OptimizerResult
 from .predict import PredictionConditionResult, PredictionResult
 from .profile import ProfileResult, ProfilerResult
 from .result import Result