+++ test/base/test_objective.py
@@ -2,7 +2,6 @@
 
 import copy
 import numbers
-import sys
 from functools import partial
 
 import numpy as np
@@ -12,11 +11,6 @@ import sympy as sp
 import pypesto
 
 from ..util import CRProblem, poly_for_sensi, rosen_for_sensi
-
-pytest_skip_aesara = pytest.mark.skipif(
-    sys.version_info >= (3, 12),
-    reason="Skipped Aesara tests on Python 3.12 or higher",
-)
 
 
 @pytest.fixture(params=[True, False])
@@ -184,46 +178,9 @@ def test_finite_difference_checks():
     )
 
 
-@pytest_skip_aesara
-def test_aesara(max_sensi_order, integrated):
-    """Test function composition and gradient computation via aesara"""
-    import aesara.tensor as aet
-
-    from pypesto.objective.aesara import AesaraObjective
-
-    prob = rosen_for_sensi(max_sensi_order, integrated, [0, 1])
-
-    # create aesara specific symbolic tensor variables
-    x = aet.specify_shape(aet.vector("x"), (2,))
-
-    # apply inverse transform such that we evaluate at prob['x']
-    x_ref = np.arcsinh(prob["x"])
-
-    # compose rosenbrock function with sinh transformation
-    obj = AesaraObjective(prob["obj"], x, aet.sinh(x))
-
-    # check function values and derivatives, also after copy
-    for _obj in (obj, copy.deepcopy(obj)):
-        # function value
-        assert _obj(x_ref) == prob["fval"]
-
-        # gradient
-        if max_sensi_order > 0:
-            assert np.allclose(
-                _obj(x_ref, sensi_orders=(1,)), prob["grad"] * np.cosh(x_ref)
-            )
-
-        # hessian
-        if max_sensi_order > 1:
-            assert np.allclose(
-                prob["hess"] * (np.diag(np.power(np.cosh(x_ref), 2)))
-                + np.diag(prob["grad"] * np.sinh(x_ref)),
-                _obj(x_ref, sensi_orders=(2,)),
-            )
-
-
 @pytest.mark.parametrize("enable_x64", [True, False])
-def test_jax(max_sensi_order, integrated, enable_x64):
+@pytest.mark.parametrize("fix_parameters", [True, False])
+def test_jax(max_sensi_order, integrated, enable_x64, fix_parameters):
     """Test function composition and gradient computation via jax"""
     import jax
     import jax.numpy as jnp
@@ -234,6 +191,7 @@ def test_jax(max_sensi_order, integrated, enable_x64):
     jax.config.update("jax_enable_x64", enable_x64)
 
     from pypesto.objective.jax import JaxObjective
+    from pypesto.objective.pre_post_process import FixedParametersProcessor
 
     prob = rosen_for_sensi(max_sensi_order, integrated, [0, 1])
 
@@ -250,9 +208,20 @@ def test_jax(max_sensi_order, integrated, enable_x64):
     # compose rosenbrock function with sinh transformation
     obj = JaxObjective(prob["obj"])
 
+    if fix_parameters:
+        obj.pre_post_processor = FixedParametersProcessor(
+            dim_full=2,
+            x_free_indices=[0],
+            x_fixed_indices=[1],
+            x_fixed_vals=[0.0],
+        )
+
     # evaluate for a couple of random points such that we can assess
     # compatibility with vmap
     xx = x_ref + np.random.randn(10, x_ref.shape[0])
+    if fix_parameters:
+        xx = xx[:, obj.pre_post_processor.x_free_indices]
+
     rvals_ref = [
         jax_op_out(
             prob["obj"](jax_op_in(xxi), sensi_orders=(max_sensi_order,))
@@ -262,6 +231,9 @@ def test_jax(max_sensi_order, integrated, enable_x64):
 
     def _fun(y, pypesto_fun, jax_fun_in, jax_fun_out):
         return jax_fun_out(pypesto_fun(jax_fun_in(y)))
+
+    assert obj.check_sensi_orders((max_sensi_order,), pypesto.C.MODE_FUN)
+    assert not obj.check_sensi_orders((max_sensi_order,), pypesto.C.MODE_RES)
 
     for _obj in (obj, copy.deepcopy(obj)):
         fun = partial(
@@ -273,6 +245,7 @@ def test_jax(max_sensi_order, integrated, enable_x64):
 
         if max_sensi_order == 1:
             fun = jax.grad(fun)
+
         # check compatibility with vmap and jit
         vmapped_fun = jax.vmap(fun)
         rvals_jax = vmapped_fun(xx)
@@ -281,8 +254,11 @@ def test_jax(max_sensi_order, integrated, enable_x64):
         # can't use rtol = 1e-8 for 32bit
         rtol = 1e-16 if enable_x64 else 1e-4
         for x, rref, rj in zip(xx, rvals_ref, rvals_jax):
+            assert isinstance(rj, jnp.ndarray)
             if max_sensi_order == 0:
-                np.testing.assert_allclose(rref, rj, atol=atol, rtol=rtol)
+                np.testing.assert_allclose(
+                    rref, float(rj), atol=atol, rtol=rtol
+                )
             if max_sensi_order == 1:
                 # g(x) = b(c(x)) => g'(x) = b'(c(x))) * c'(x)
                 # f(x) = a(g(x)) => f'(x) = a'(g(x)) * g'(x)
@@ -295,7 +271,9 @@ def test_jax(max_sensi_order, integrated, enable_x64):
                 ) @ jax.jacfwd(jax_op_in)(x)
                 # f'(x) = a'(g(x)) * g'(x)
                 f_prime = jax.jacfwd(jax_op_out)(g) * g_prime
-                np.testing.assert_allclose(f_prime, rj, atol=atol, rtol=rtol)
+                np.testing.assert_allclose(
+                    f_prime, np.asarray(rj), atol=atol, rtol=rtol
+                )
 
 
 @pytest.fixture(
@@ -320,12 +298,20 @@ def fd_delta(request):
     return request.param
 
 
-def test_fds(fd_method, fd_delta):
+# add a fixture for fixed and unfixed parameters
+@pytest.mark.parametrize("fixed", [True, False])
+def test_fds(fd_method, fd_delta, fixed):
     """Test finite differences."""
     problem = CRProblem()
 
-    # reference objective
-    obj = problem.get_objective()
+    if fixed:
+        fixed_problem = problem.get_problem()
+        fixed_problem.fix_parameters([1], problem.p_true[1])
+        obj = fixed_problem.objective
+        p = problem.p_true[0]
+    else:
+        obj = problem.get_objective()
+        p = problem.p_true
 
     # FDs for everything
     obj_fd = pypesto.FD(
@@ -372,7 +358,6 @@ def test_fds(fd_method, fd_delta):
         delta_grad=fd_delta,
         delta_res=fd_delta,
     )
-    p = problem.p_true
 
     # check that function values coincide (call delegated)
     for attr in ["fval", "res"]:
@@ -425,3 +410,29 @@ def test_fds(fd_method, fd_delta):
         assert obj_fd.delta_fun.updates == 0
     else:
         assert obj_fd.delta_fun.updates > 1
+
+
+def test_call_unprocessed_return_dict():
+    class Objective0(pypesto.objective.ObjectiveBase):
+        def call_unprocessed(self, *args, **kwargs):
+            return {"fval": 0, "return_dict": "return_dict" in kwargs}
+
+        def check_sensi_orders(self, *args, **kwargs):
+            return True
+
+    class Objective1(Objective0):
+        def call_unprocessed(self, *args, return_dict: bool, **kwargs):
+            return {"fval": 0, "return_dict": return_dict}
+
+    objective0 = Objective0()
+    objective1 = Objective1()
+
+    with pytest.warns(DeprecationWarning, match="Please add `return_dict`"):
+        result0 = objective0([0], return_dict=True)
+
+    result1 = objective1([0], return_dict=True)
+
+    # `return_dict` is not shared with `call_unprocessed` by default,
+    assert not result0["return_dict"]
+    # but is shared if the `call_unprocessed` signature supports it.
+    assert result1["return_dict"]