+++ test/base/test_roadrunner.py
@@ -4,11 +4,14 @@ import logging
 import os
 
 import benchmark_models_petab as models
-import petab
+import numpy as np
+import petab.v1 as petab
 import petabtests
 import pytest
 
-import pypesto.objective.roadrunner as objective_rr
+import pypesto
+import pypesto.petab
+from pypesto.objective.roadrunner import simulation_to_measurement_df
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -58,19 +61,21 @@ def _execute_case_rr(case, model_type, version):
     # import petab problem
     yaml_file = case_dir / petabtests.problem_yaml_name(case)
 
-    importer = objective_rr.PetabImporterRR.from_yaml(yaml_file)
+    importer = pypesto.petab.PetabImporter.from_yaml(
+        yaml_file, simulator_type="roadrunner"
+    )
     petab_problem = importer.petab_problem
-    obj = importer.create_objective()
+    obj = importer.create_problem().objective
 
     # the scaled parameters
-    problem_parameters = importer.petab_problem.x_nominal_scaled
+    problem_parameters = importer.petab_problem.x_nominal_free_scaled
 
     # simulate
     ret = obj(problem_parameters, sensi_orders=(0,), return_dict=True)
 
     # extract results
     llh = -ret["fval"]
-    simulation_df = objective_rr.simulation_to_measurement_df(
+    simulation_df = simulation_to_measurement_df(
         ret["simulation_results"], petab_problem.measurement_df
     )
 
@@ -115,7 +120,9 @@ def test_deepcopy():
         os.path.join(models.MODELS_DIR, model_name, model_name + ".yaml")
     )
     petab_problem.model_name = model_name
-    importer = objective_rr.PetabImporterRR(petab_problem)
+    importer = pypesto.petab.PetabImporter(
+        petab_problem, simulator_type="roadrunner"
+    )
     problem_parameters = petab_problem.x_nominal_free_scaled
 
     problem = importer.create_problem()
@@ -136,3 +143,43 @@ def test_deepcopy():
     )
 
     assert obj(problem_parameters) != copied_objective(problem_parameters)
+
+
+def test_multiprocessing():
+    """Test that multiprocessing works as intended"""
+    model_name = "Boehm_JProteomeRes2014"
+    petab_problem = petab.Problem.from_yaml(
+        os.path.join(models.MODELS_DIR, model_name, model_name + ".yaml")
+    )
+    petab_problem.model_name = model_name
+    importer = pypesto.petab.PetabImporter(
+        petab_problem, simulator_type="roadrunner"
+    )
+
+    problem = importer.create_problem()
+    # start 30 times from the same point
+    start_points = [problem.get_full_vector(problem.get_startpoints(1))] * 30
+    problem.set_x_guesses(np.vstack(start_points))
+
+    # for later comparisons, do one optimization run with single core
+    result_single = pypesto.optimize.minimize(
+        problem=problem,
+        n_starts=1,
+        engine=pypesto.engine.SingleCoreEngine(),
+        progress_bar=False,
+    )
+
+    engine = pypesto.engine.MultiProcessEngine(n_procs=8)
+
+    result = pypesto.optimize.minimize(
+        problem=problem,
+        n_starts=15,
+        engine=engine,
+        progress_bar=True,
+    )
+    assert np.all(
+        [
+            fval == result_single.optimize_result.fval[0]
+            for fval in result.optimize_result.fval
+        ]
+    )